+++ apps/fabric-website-resources/src/components/pages/DetailsList/DetailsListPage.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { DemoPage } from '../DemoPage';
+import { DemoPage } from '../../DemoPage';
 import { DetailsListPageProps } from 'office-ui-fabric-react/lib/components/DetailsList/DetailsList.doc';
 
 export const DetailsListPage = (props: { isHeaderVisible: boolean }) => (