+++ packages/office-ui-fabric-react/src/components/Check/Check.base.tsx
@@ -1,11 +1,10 @@
 import * as React from 'react';
-import { BaseComponent, customizable } from '../../Utilities';
+import { BaseComponent } from '../../Utilities';
 import { ICheckProps } from './Check.types';
 import { Icon } from '../../Icon';
 import { getStyles } from './Check.styles';
 import { getClassNames } from './Check.classNames';
 
-@customizable('Check', ['theme', 'styles'])
 export class CheckBase extends BaseComponent<ICheckProps, {}> {
   public static defaultProps: ICheckProps = {
     checked: false