+++ packages/office-ui-fabric-react/src/components/Spinner/Spinner.types.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { ITheme, IStyle } from '../../Styling';
-import { IStyleFunctionOrObject } from '../../Utilities';
+import { IRefObject, IStyleFunctionOrObject } from '../../Utilities';
 
 export interface ISpinner {}
 
@@ -9,7 +9,7 @@ export interface ISpinnerProps extends React.HTMLAttributes<HTMLElement> {
    * Optional callback to access the ISpinner interface. Use this instead of ref for accessing
    * the public methods and properties of the component.
    */
-  componentRef?: (component: ISpinner | null) => void;
+  componentRef?: IRefObject<ISpinner>;
 
   /**
    * Deprecated and will be removed at >= 2.0.0. Use SpinnerSize instead.