+++ packages/example-app-base/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log - @uifabric/example-app-base
 
-This log was last generated on Fri, 13 Jul 2018 21:32:37 GMT and should not be manually modified.
+This log was last generated on Mon, 16 Jul 2018 10:27:18 GMT and should not be manually modified.
+
+## 6.4.0
+Mon, 16 Jul 2018 10:27:18 GMT
+
+### Minor changes
+
+- Variant section lists have same style as best practices section lists. ComponentPage Overview now marked optional prop.
+- EditSection: Remove ComponentPageSection enum to make EditSection easier to use.
 
 ## 6.3.0
 Fri, 13 Jul 2018 21:32:37 GMT